// src/libs/common/enums/date-preset.enum.ts
export enum DatePreset {
  Today = 'today',
  Yesterday = 'yesterday',
  ThisWeek = 'thisWeek',
  Last7Days = 'last7Days',
  LastWeek = 'lastWeek',
  Last28Days = 'last28Days',
  Last30Days = 'last30Days',
  ThisMonth = 'thisMonth',
  LastMonth = 'lastMonth',
  Last90Days = 'last90Days',
  Last3Months = 'last3Months',
  QuarterToDate = 'quarterToDate',
  LastQuarter = 'lastQuarter',
  Last6Months = 'last6Months',
  Last12Months = 'last12Months',
  ThisYear = 'thisYear',
  LastCalendarYear = 'lastCalendarYear',
}